Implement `WasmBindgen` executor and reorganize
