+++ futures/Cargo.toml
@@ -27,3 +27,6 @@ features = ["rt-core"]
 [dependencies.async-std]
 version = "1.0"
 optional = true
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen-futures = "0.4"
