--- src/web.rs
@@ -1,1 +1,0 @@
-pub use iced_web::*;
